requirements.txt
plaid-python
gspread
protobuf
python-dotenv
